quest-generator/requirements.txt
Flask>=2.0
groq>=0.4.0
python-dotenv>=1.0.0
llama-cpp-python>=0.2.0
pytest>=7.0
pytest-mock>=3.0
